Moved flow run listeners from device to driver
