+++ drivers/zappi/device.ts
@@ -92,88 +92,6 @@ export class ZappiDevice extends Device {
       dev.InitializeCapabilities();
     });
 
-    // Flow logic
-    const chargingCondition = dev.homey.flow.getConditionCard('is_charging');
-    chargingCondition.registerRunListener(async (args, state) => {
-      dev.log(`Is Charging: ${args} - ${state}`);
-      const charging = dev._chargerStatus === ZappiStatus.Charging; // true or false
-      return charging;
-    });
-
-    const startChargingAction = dev.homey.flow.getActionCard('start_charging');
-    startChargingAction.registerRunListener(async (args, state) => {
-      dev.log(`Start Charging: ${args} - ${state}`);
-      try {
-        await dev.setChargerState(true);
-      } catch (error) {
-        dev.error(error);
-      }
-    });
-
-    const stopChargingAction = dev.homey.flow.getActionCard('stop_charging');
-    stopChargingAction.registerRunListener(async (args, state) => {
-      dev.log(`Stop Charging: ${args} - ${state}`);
-      try {
-        await dev.setChargerState(false);
-      } catch (error) {
-        dev.error(error);
-      }
-    });
-
-    const setChargeModeAction = dev.homey.flow.getActionCard('set_charge_mode');
-    setChargeModeAction.registerRunListener(async (args, state) => {
-      dev.log(`Charge Mode: ${args.charge_mode_txt}`);
-      try {
-        dev._chargeMode = dev.getChargeMode(args.charge_mode_txt);
-        if (dev._chargeMode !== ZappiChargeMode.Off) {
-          dev._lastOnState = dev._chargeMode;
-        }
-        await dev.setChargeMode(dev._chargeMode);
-      } catch (error) {
-        dev.error(error);
-      }
-    });
-
-    const selectChargeModeAction = dev.homey.flow.getActionCard('select_charge_mode');
-    selectChargeModeAction.registerRunListener(async (args, state) => {
-      dev.log(`Charge Mode: ${args.charge_mode_selector}`);
-      try {
-        dev._chargeMode = dev.getChargeMode(args.charge_mode_selector);
-        if (dev._chargeMode !== ZappiChargeMode.Off) {
-          dev._lastOnState = dev._chargeMode;
-        }
-        await dev.setChargeMode(dev._chargeMode);
-      } catch (error) {
-        dev.error(error);
-      }
-    });
-
-    const setBoostModeAction = dev.homey.flow.getActionCard('set_boost_mode');
-    setBoostModeAction.registerRunListener(async (args, state) => {
-      dev.log(`Boost Mode: ${args.boost_mode_txt}, Boost Mode: ${args.boost_mode_kwh}, Boost Mode: ${args.boost_mode_complete_time}`);
-      const kwh = args.boost_mode_kwh ? args.boost_mode_kwh as number : 0;
-      const completeTime = dev.getValidBoostTime(args.boost_mode_complete_time ? args.boost_mode_complete_time : '0000');
-      dev.log(`Complete time: ${completeTime}`);
-      try {
-        dev._boostMode = dev.getBoostMode(args.boost_mode_txt);
-        dev._lastBoostState = dev._boostMode;
-        await dev.setBoostMode(dev._boostMode, kwh, completeTime);
-      } catch (error) {
-        dev.error(error);
-      }
-    });
-
-    const setMinimumGreenLevelAction = dev.homey.flow.getActionCard('set_minimum_green_level');
-    setMinimumGreenLevelAction.registerRunListener(async (args, state) => {
-      dev.log(`Minimum Green Level: ${args.minimum_green_level}`);
-      dev._minimumGreenLevel = args.minimum_green_level;
-      try {
-        await dev.setMinimumGreenLevel(dev._minimumGreenLevel);
-      } catch (error) {
-        dev.error(error);
-      }
-    });
-
     dev.log(`ZappiDevice ${dev.deviceId} has been initialized`);
   }
 
@@ -592,8 +510,7 @@ export class ZappiDevice extends Device {
       dev.triggerChargingFlow(isOn);
       dev.log(`Zappi was switched ${isOn ? 'on' : 'off'}`);
     } catch (error) {
-      dev.error(error);
-      throw new Error(`Switching the Zappi ${isOn ? 'on' : 'off'} failed!`);
+      dev.error(`Switching the Zappi ${isOn ? 'on' : 'off'} failed:\n${error}`);
     }
   }
 
@@ -611,8 +528,7 @@ export class ZappiDevice extends Device {
       dev.setCapabilityValue('minimum_green_level', value).catch(dev.error);
       dev.setCapabilityValue('set_minimum_green_level', value).catch(dev.error);
     } catch (error) {
-      dev.error(error);
-      throw new Error(`Switching the Zappi ${value ? 'on' : 'off'} failed!`);
+      dev.error(`Setting minimum greenn level to ${value} failed:\n${error}`);
     }
   }
 
@@ -637,8 +553,7 @@ export class ZappiDevice extends Device {
       dev.triggerChargeModeFlow(chargeMode);
       dev.log(`Zappi changed charge mode ${dev.getChargeModeText(chargeMode)}`);
     } catch (error) {
-      dev.error(error);
-      throw new Error(`Switching the Zappi charge mode ${dev.getChargeModeText(chargeMode)} failed!`);
+      dev.error(`Switching the Zappi charge mode ${dev.getChargeModeText(chargeMode)} failed:\n${error}`);
     }
   }
 
@@ -662,8 +577,7 @@ export class ZappiDevice extends Device {
       dev.triggerBoostModeFlow(boostMode);
       dev.log(`Zappi changed boost mode ${dev.getBoostModeText(boostMode)}`);
     } catch (error) {
-      dev.error(error);
-      throw new Error(`Switching the Zappi charge mode ${(dev.getBoostModeText(boostMode))} failed!`);
+      dev.error(`Switching the Zappi boost mode ${(dev.getBoostModeText(boostMode))} failed:\n${error}`);
     }
   }
